[WIP] Remove the strategy package
