--- src/torchjd/autojac/_transform/strategy/__init__.py
@@ -1,1 +1,0 @@
-from .unifying import UnifyingStrategy
