+++ src/torchjd/autojac/backward.py
@@ -5,7 +5,6 @@ from torch import Tensor
 from torchjd.aggregation import Aggregator
 
 from ._transform import Diagonalize, EmptyTensorDict, Init, Jac, Store, make_aggregation
-from ._transform.strategy import UnifyingStrategy
 from ._utils import (
     _as_tensor_list,
     _check_optional_positive_chunk_size,
@@ -81,7 +80,7 @@ def backward(
     jac = Jac(tensors, inputs, parallel_chunk_size, retain_graph)
 
     # Transform that aggregates the Jacobians.
-    aggregate = make_aggregation(UnifyingStrategy(A, inputs))
+    aggregate = make_aggregation(A, inputs)
 
     # Transform that stores the result in the .grad field of the inputs.
     store = Store(inputs)
